# lambda/server/openapi.yaml
openapi: 3.0.3
info:
  title: MarginArc Engine API
  description: |
    The MarginArc Engine API provides margin recommendation, BOM optimization,
    deal outcome recording, and product catalog search for Value-Added Resellers (VARs).

    All margin values use the **margin-on-selling-price** convention:
    `marginPct = (price − cost) / price`.
  version: 1.0.0
  contact:
    name: MarginArc Support
    url: https://marginarc.com
servers:
  - url: https://api.marginarc.com
    description: Production
security:
  - ApiKeyAuth: []
components:
  securitySchemes:
    ApiKeyAuth:
      type: apiKey
      in: header
      name: x-api-key
      description: API key provided during onboarding. Set in Salesforce custom setting `Fulcrum_Config__c.API_Key__c`.
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: Admin JWT token obtained from `/admin/api/auth/login`.
  schemas:
...
    ProductCategory:
      type: string
      enum:
        - Hardware
        - Software
        - Cloud
        - ProfessionalServices
        - ManagedServices
        - ComplexSolution
    CustomerSegment:
      type: string
      enum: [SMB, MidMarket, Enterprise]
    RelationshipStrength:
      type: string
      enum: [New, Good, Strategic]
    CustomerTechSophistication:
      type: string
      enum: [Low, Medium, High]
    DealRegType:
      type: string
      enum: [NotRegistered, StandardApproved, PremiumHunting, Teaming]
    Competitors:
      type: string
      enum: ["0", "1", "2", "3+"]
    ValueAdd:
      type: string
      enum: [Low, Medium, High]
    SolutionComplexity:
      type: string
      enum: [Low, Medium, High]
    VarStrategicImportance:
      type: string
      enum: [Low, Medium, High]
    CustomerIndustry:
      type: string
      description: Industry vertical of the end customer.
      enum:
        - Consumer Goods & Food
        - Diversified Conglomerates
        - Energy
        - Financial Services
        - Life Sciences & Healthcare
        - Manufacturing & Automotive
        - Media & Telecommunications
        - Retail
        - Technology
        - Transportation & Logistics
...
    OemProfile:
      type: object
      description: Optional OEM-specific margin parameters configured in admin portal.
      properties:
        baseMargin:
          type: number
          minimum: 0
          maximum: 100
          description: Base margin percentage for this OEM.
        dealRegBoost:
          type: number
          minimum: 0
          maximum: 50
          description: Additional margin points when deal is registered.
        quarterEndDiscount:
          type: number
          minimum: 0
          maximum: 50
          description: Typical OEM quarter-end discount percentage.
        servicesBoost:
          type: number
          minimum: 0
          maximum: 50
          description: Additional margin when services are attached.
        productCategory:
          type: string
          description: Primary product category for this OEM.
      nullable: true
    CompetitorProfile:
      type: object
      description: Competitor intelligence used to adjust recommendations.
      properties:
        name:
          type: string
          description: Competitor company name.
        priceAggression:
          type: number
          minimum: 1
          maximum: 5
          description: How aggressively the competitor prices (1=passive, 5=very aggressive).
        marginAggression:
          type: number
          minimum: -5
          maximum: 5
          description: Margin aggression factor.
        typicalDiscount:
          type: number
          minimum: 0
          maximum: 100
          description: Typical discount percentage offered by this competitor.
        servicesCapability:
          type: number
          minimum: 1
          maximum: 5
          description: Competitor's services capability (1=weak, 5=strong).
        primaryOems:
          type: string
          description: Primary OEM partnerships.
        primaryStrength:
          type: string
          description: Competitor's primary competitive advantage.
      required: [name]
    DealInput:
      type: object
      description: Deal parameters used by the recommendation engine.
      required:
        - oemCost
        - productCategory
        - customerSegment
        - customerIndustry
      properties:
        oem:
          type: string
          description: OEM vendor name (e.g. "Cisco", "Dell Technologies").
          example: Cisco
        oemCost:
          type: number
          description: Total OEM cost (buy price) in dollars. Must be positive.
          example: 100000
        productCategory:
          $ref: '#/components/schemas/ProductCategory'
        customerSegment:
          $ref: '#/components/schemas/CustomerSegment'
        relationshipStrength:
          allOf:
            - $ref: '#/components/schemas/RelationshipStrength'
          default: Good
          description: Defaults to "Good" if omitted.
        customerTechSophistication:
          allOf:
            - $ref: '#/components/schemas/CustomerTechSophistication'
          default: Medium
          description: Defaults to "Medium" if omitted.
        dealRegType:
          allOf:
            - $ref: '#/components/schemas/DealRegType'
          default: NotRegistered
          description: Defaults to "NotRegistered" if omitted.
        competitors:
          allOf:
            - $ref: '#/components/schemas/Competitors'
          default: "1"
          description: Defaults to "1" if omitted.
        valueAdd:
          allOf:
            - $ref: '#/components/schemas/ValueAdd'
          default: Medium
          description: Defaults to "Medium" if omitted.
        solutionComplexity:
          allOf:
            - $ref: '#/components/schemas/SolutionComplexity'
          default: Medium
          description: Defaults to "Medium" if omitted.
        varStrategicImportance:
          allOf:
            - $ref: '#/components/schemas/VarStrategicImportance'
          default: Medium
          description: Defaults to "Medium" if omitted.
        customerIndustry:
          $ref: '#/components/schemas/CustomerIndustry'
        customerPriceSensitivity:
          type: integer
          minimum: 1
          maximum: 5
          description: Customer's price sensitivity (1=not sensitive, 5=very sensitive).
        customerLoyalty:
          type: integer
          minimum: 1
          maximum: 5
          description: Customer loyalty to your company (1=low, 5=very loyal).
        dealUrgency:
          type: integer
          minimum: 1
          maximum: 5
          description: How urgently the customer needs to close (1=no rush, 5=critical).
        isNewLogo:
          type: boolean
          description: Whether this is a new customer (true) or existing (false).
        solutionDifferentiation:
          type: integer
          minimum: 1
          maximum: 5
          description: How differentiated your solution is (1=commodity, 5=unique).
        servicesAttached:
          type: boolean
          description: Whether professional/managed services are attached to this deal.
        quarterEnd:
          type: boolean
          description: Whether the deal is closing at OEM quarter-end (may unlock extra discounts).
        displacementDeal:
          type: boolean
          description: Whether this deal displaces an incumbent competitor.
        dealSize:
          type: number
          description: Total deal size in dollars (optional, informational).
        dealType:
          type: string
          description: Free-text deal type label.
        accountName:
          type: string
          description: Customer account name.
        competitorNames:
          type: array
          items:
            type: string
          description: List of competing VAR names on this deal.
          example: ["CDW", "SHI"]
        oemProfile:
          $ref: '#/components/schemas/OemProfile'
        competitorProfiles:
          type: array
          items:
            $ref: '#/components/schemas/CompetitorProfile'
          nullable: true
          description: Detailed competitor profiles for named competitors.
    BomLineInput:
      type: object
      description: A single BOM line item submitted with a deal recommendation request.
      required:
        - description
        - listPrice
        - discountedPrice
        - priceAfterMargin
      properties:
        description:
          type: string
          description: Line item description.
          example: Catalyst 9300 48-Port PoE+ Switch
        key:
          type: string
          description: Optional unique key for the line.
        category:
          type: string
          default: Hardware
          description: Product category for this line.
        unit:
          type: string
          default: ea
          description: Unit of measure.
        productNumber:
          type: string
          description: Vendor part number.
          example: C9300-48P-A
        productId:
          type: string
          description: Internal product identifier.
        vendor:
          type: string
          description: Vendor/manufacturer name.
          example: Cisco
        listPrice:
          type: number
          minimum: 0
          description: List (MSRP) price per unit.
          example: 12495
        discountedPrice:
          type: number
          minimum: 0
          description: Your cost after OEM discount.
          example: 8746.50
        priceAfterMargin:
          type: number
          minimum: 0
          description: Sell price to customer (cost + margin).
          example: 10295.88
        recommendedMarginPct:
          type: number
          minimum: 0
          maximum: 100
          description: Recommended margin percentage for this line.
        quantity:
          type: number
          minimum: 0.01
          default: 1
          description: Quantity of this line item.
          example: 10
        note:
          type: string
          description: Optional note for this line.

    # ── Response schemas ───────────────────────────────────────────────────────
    Driver:
      type: object
      properties:
        factor:
          type: string
          description: Name of the margin driver.
          example: OEM Profile
        direction:
          type: string
          enum: [up, down]
          description: Whether this factor pushes margin up or down.
        impact:
          type: number
          description: Magnitude of impact in margin percentage points.
          example: 2.5
        explanation:
          type: string
          description: Human-readable explanation of this driver.
    PredictionQuality:
      type: object
      description: Data quality assessment of the input fields.
      properties:
        score:
          type: integer
          minimum: 0
          maximum: 100
          description: Overall quality score (0–100).
          example: 72
        grade:
          type: string
          enum: [Excellent, Good, Fair, Poor]
          description: Quality grade based on score.
          example: Good
        missingFields:
          type: array
          items:
            type: string
          description: Up to 5 optional fields that would improve prediction quality.
          example: ["Price sensitivity (1-5)", "Deal urgency (1-5)"]
    ScoreFactors:
      type: object
      description: Breakdown of the deal score components.
      properties:
        marginAlignment:
          type: object
          properties:
            score:
              type: integer
              description: Points earned for margin alignment.
            max:
              type: integer
              description: Maximum possible points.
              example: 40
        winProbability:
          type: object
          properties:
            score:
              type: integer
            max:
              type: integer
              example: 25
        dataQuality:
          type: object
          properties:
            score:
              type: integer
            max:
              type: integer
              example: 20
        algorithmConfidence:
          type: object
          properties:
            score:
              type: integer
            max:
              type: integer
              example: 15
    Metrics:
      type: object
      description: Financial comparison between planned and recommended margins.
      properties:
        planned:
          type: object
          properties:
            marginPct:
              type: number
            price:
              type: number
            grossProfit:
              type: number
            riskAdjusted:
              type: number
        recommended:
          type: object
          properties:
            marginPct:
              type: number
            price:
              type: number
            grossProfit:
              type: number
            riskAdjusted:
              type: number
        delta:
          type: object
          properties:
            marginPct:
              type: number
            price:
              type: number
            grossProfit:
              type: number
            riskAdjusted:
              type: number
    RecommendResponse:
      type: object
      properties:
        suggestedMarginPct:
          type: number
          nullable: true
          description: Recommended margin percentage (margin-on-selling-price). Null in Phase 1.
          example: 18.5
        suggestedPrice:
          type: number
          nullable: true
          description: Recommended sell price based on suggestedMarginPct. Null in Phase 1.
          example: 122699.39
        algorithmMarginPct:
          type: number
          description: Raw algorithm margin before any BOM override.
          example: 18.5
        algorithmSuggestedPrice:
          type: number
          description: Raw algorithm price before any BOM override.
        winProbability:
          type: number
          description: Estimated win probability (0–1).
          example: 0.72
        confidence:
          type: number
          description: Algorithm confidence in this recommendation (0–1).
          example: 0.85
        method:
          type: string
          description: Method used to compute margin (e.g. "kNN + rules", "Manual BOM blend").
          example: kNN + rules
        drivers:
          type: array
          items:
            $ref: '#/components/schemas/Driver'
          description: Factors that influenced the margin recommendation.
        policyFloor:
          type: number
          description: Minimum margin enforced by policy guardrails.
          example: 5
        explanation:
          type: string
          description: AI-generated natural language explanation of the recommendation.
        qualitativeSummary:
          type: string
          description: AI-generated qualitative deal narrative.
        metrics:
          $ref: '#/components/schemas/Metrics'
        bom:
          type: object
          description: Synthesized or manual Bill of Materials. Lines included in Phase 3 only.
          properties:
            origin:
              type: string
              enum: [synthesized, manual]
            totals:
              type: object
              properties:
                price:
                  type: number
                marginPct:
                  type: number
            stats:
              type: object
              properties:
                lineCount:
                  type: integer
            lines:
              type: array
              items:
                type: object
              description: Detailed BOM lines (Phase 3 only).
        predictionQuality:
          $ref: '#/components/schemas/PredictionQuality'
        dealScore:
          type: integer
          minimum: 0
          maximum: 100
          description: Composite deal score (0–100).
          example: 74
        scoreFactors:
          $ref: '#/components/schemas/ScoreFactors'
        phaseInfo:
          type: object
          description: Customer phase information.
          properties:
            current:
              type: integer
              enum: [1, 2, 3]
              description: Current algorithm phase.
            message:
              type: string
              description: Phase guidance message (Phase 1 only).
            nextPhaseReady:
              type: boolean
              description: Whether the customer is ready for the next phase (Phase 1 only).

    # ── BOM Search ─────────────────────────────────────────────────────────────
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
          description: Human-readable error message.
          example: Invalid or missing API key
paths:
  /api/recommend:
    post:
      operationId: recommend
      summary: Get margin recommendation
      description: |
        Core recommendation endpoint. Accepts deal parameters (and optionally a planned
        margin and BOM lines), returns a recommended margin percentage, win probability,
        confidence score, margin drivers, financial metrics, and an AI-generated explanation.

        The response varies by customer phase:
        - **Phase 1 (Score Only)**: Returns `dealScore` and `scoreFactors` but `suggestedMarginPct` is null.
        - **Phase 2 (Recommend)**: Full recommendation without detailed BOM lines.
        - **Phase 3 (Optimize)**: Full recommendation including detailed BOM lines.
      tags: [Recommendations]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [input]
              properties:
                input:
                  $ref: '#/components/schemas/DealInput'
                plannedMarginPct:
                  type: number
                  nullable: true
                  description: The sales rep's planned margin percentage. Used for comparison metrics.
                  example: 15
                bomLines:
                  type: array
                  items:
                    $ref: '#/components/schemas/BomLineInput'
                  description: Optional manual BOM lines. When provided, the blended BOM margin overrides the algorithm margin.
                  maxItems: 50
            examples:
              minimal:
                summary: Minimal payload (only required fields)
                value:
                  input:
                    oemCost: 100000
                    productCategory: Hardware
                    customerSegment: MidMarket
                    customerIndustry: Technology
                  plannedMarginPct: 15
              full:
                summary: Full payload with all optional fields
                value:
                  input:
                    oem: Cisco
                    oemCost: 100000
                    productCategory: Hardware
                    customerSegment: MidMarket
                    dealRegType: StandardApproved
                    competitors: "1"
                    competitorNames: ["CDW"]
                    valueAdd: High
                    solutionComplexity: Medium
                    relationshipStrength: Good
                    customerIndustry: Technology
                    customerTechSophistication: Medium
                    servicesAttached: false
                    quarterEnd: false
                    isNewLogo: false
                    varStrategicImportance: Medium
                  plannedMarginPct: 15
                  bomLines: []
      responses:
        "200":
          description: Recommendation computed successfully.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RecommendResponse'
              example:
                suggestedMarginPct: 18.5
                suggestedPrice: 122699.39
                algorithmMarginPct: 18.5
                algorithmSuggestedPrice: 122699.39
                winProbability: 0.72
                confidence: 0.85
                method: "kNN + rules"
                drivers:
                  - factor: OEM Profile
                    direction: up
                    impact: 2.5
                    explanation: "Cisco standard approved deal registration unlocks higher margin"
                policyFloor: 5
                explanation: "Margin set at 18.5%. Key contributors: Deal Registration 4.0%, Relationship 2.0%."
                qualitativeSummary: "With 1 competitor, a good relationship, and standard approved registration, an 18.5% blend balances win odds and profitability."
                metrics:
                  planned:
                    marginPct: 15
                    price: 117647.06
                    grossProfit: 17647.06
                    riskAdjusted: 12705.88
                  recommended:
                    marginPct: 18.5
                    price: 122699.39
                    grossProfit: 22699.39
                    riskAdjusted: 16343.56
                  delta:
                    marginPct: 3.5
                    price: 5052.33
                    grossProfit: 5052.33
                    riskAdjusted: 3637.68
                predictionQuality:
                  score: 72
                  grade: Good
                  missingFields:
                    - "Price sensitivity (1-5)"
                    - "Deal urgency (1-5)"
                dealScore: 74
                scoreFactors:
                  marginAlignment:
                    score: 26
                    max: 40
                  winProbability:
                    score: 18
                    max: 25
                  dataQuality:
                    score: 14
                    max: 20
                  algorithmConfidence:
                    score: 13
                    max: 15
                phaseInfo:
                  current: 2
        "400":
          description: Invalid input (validation error).
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              example:
                error: "Expected 'Hardware' | 'Software' | 'Cloud' | 'ProfessionalServices' | 'ManagedServices' | 'ComplexSolution'"
        "401":
          description: Missing or invalid API key.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              example:
                error: Invalid or missing API key
...